Update governance.md
First draft of governance page. Might need to tweak it a little bit later on. @roll have a look and see what you think. Mainly: do we want to mention wg members are part of the Frictionless community? It's nice to keep the link ATM, but we might want to skip that if the wg changes composition in the future.