Add support for labeled missingness
Resolves https://github.com/frictionlessdata/specs/issues/880. When
paired with the `categorical` type, this gives us full support for the
value labels found in many statistical software packages.

I have included it here rather than a separate PR because these
issues are intertwined and there's a synergy in addressing them
simultaneously. That said, if you would rather see this in a PR, let me
know and I can revert.